+++ tests/Phan/Language/UnionTypeTest.php
@@ -156,6 +156,7 @@ class UnionTypeTest extends BaseTest
         // TODO: Implement more code to warn about invalid operands.
         // Evaluating this should result in '0'
         $this->assertUnionTypeStringEqual('$x=(new stdClass()); $x &= 2; $x', 'int');
+        $this->assertUnionTypeStringEqual('$x = stdClass::class; new $x();', '\stdClass');
     }
 
     public function testString()
@@ -240,6 +241,7 @@ class UnionTypeTest extends BaseTest
     public function testGenericArrayTypeFromString()
     {
         $type = Type::fromFullyQualifiedString("int[][]");
+        assert($type instanceof GenericArrayType);
 
         $this->assertEquals(
             $type->genericArrayElementType()->__toString(),
@@ -304,6 +306,18 @@ class UnionTypeTest extends BaseTest
     {
         $union_type = self::makePHPDocUnionType($union_type_string);
         $this->assertTrue($union_type->hasType($type), "Expected $union_type (from $union_type_string) to be $type");
+    }
+
+    public function testExpandedTypes()
+    {
+        $this->assertSame(
+            '\Exception[]|\Throwable[]',
+            UnionType::fromFullyQualifiedString('\Exception[]')->asExpandedTypes(self::$code_base)->__toString()
+        );
+        $this->assertSame(
+            'array<int,\Exception>|array<int,\Throwable>',
+            UnionType::fromFullyQualifiedString('array<int,\Exception>')->asExpandedTypes(self::$code_base)->__toString()
+        );
     }
 
     public function testBasicTypes()